# apps/dashboard/openapi.json
{
  "components": {
    "schemas": {
      "DeployInputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/DeployInputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "image_tag": {
            "description": "Full Docker image reference to deploy (e.g. registry.example.com/app:sha256)",
            "type": "string"
          }
        },
        "required": [
          "image_tag"
        ],
        "type": "object"
      },
      "DeployOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/DeployOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "app": {
            "description": "App name derived from image_tag (last path segment before the colon)",
            "type": "string"
          },
          "error": {
            "description": "Error message present only when status is failed",
            "type": "string"
          },
          "job_id": {
            "description": "Unique job identifier (UUID)",
            "type": "string"
          },
          "status": {
            "description": "Job status: queued, running, success, or failed",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "app",
          "status"
        ],
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ErrorModel.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
      "HealthOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/HealthOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "version"
        ],
        "type": "object"
      },
      "Job": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/Job.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "app": {
            "type": "string"
          },
          "compose_path": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "image_tag": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "app",
          "status",
          "image_tag",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Webhook-driven rolling deployment orchestrator for Docker Compose stacks",
    "title": "RollHook API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/deploy": {
      "post": {
        "description": "Enqueues a zero-downtime rolling deploy for the service matching image_tag. By default blocks until the job reaches a terminal state and returns the result. Pass ?async=true to return immediately with status=queued. The app name is derived from the last path segment of image_tag before the colon (e.g. ghcr.io/org/my-api:sha → my-api). Returns 500 with an error field if the deploy fails.",
        "operationId": "post-deploy",
        "parameters": [
          {
            "description": "Return immediately with status=queued instead of blocking until completion",
            "explode": false,
            "in": "query",
            "name": "async",
            "schema": {
              "description": "Return immediately with status=queued instead of blocking until completion",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeployOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Trigger a rolling deployment",
        "tags": [
          "Deploy"
        ]
      }
    },
    "/health": {
      "get": {
        "description": "Returns server status and deployed version. Returns 503 during graceful shutdown so load balancers deregister this backend before the process exits.",
        "operationId": "get-health",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Health check",
        "tags": [
          "Health"
        ]
      }
    },
    "/jobs": {
      "get": {
        "description": "Returns deployment jobs ordered newest-first. Supports optional filters by app name and status. Default limit is 50.",
        "operationId": "get-jobs",
        "parameters": [
          {
            "description": "Filter by app name",
            "explode": false,
            "in": "query",
            "name": "app",
            "schema": {
              "description": "Filter by app name",
              "type": "string"
            }
          },
          {
            "description": "Filter by status: queued, running, success, or failed",
            "explode": false,
            "in": "query",
            "name": "status",
            "schema": {
              "description": "Filter by status: queued, running, success, or failed",
              "type": "string"
            }
          },
          {
            "description": "Maximum results to return (default 50)",
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "description": "Maximum results to return (default 50)",
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List deployment jobs",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}": {
      "get": {
        "description": "Returns a single deployment job by ID. Returns 404 if not found. compose_path and service are populated after discovery completes; error is set only when status is failed.",
        "operationId": "get-job",
        "parameters": [
          {
            "description": "Job UUID returned by POST /deploy",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Job UUID returned by POST /deploy",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get job by ID",
        "tags": [
          "Jobs"
        ]
      }
    }
  }
}