# apps/dashboard/src/api/generated/models/deployInputBody.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * RollHook API
 * Webhook-driven rolling deployment orchestrator for Docker Compose stacks
 * OpenAPI spec version: 0.1.0
 */

export interface DeployInputBody {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** Full Docker image reference to deploy (e.g. registry.example.com/app:sha256) */
  image_tag: string;
}