# apps/dashboard/src/api/generated/models/deployOutputBody.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * RollHook API
 * Webhook-driven rolling deployment orchestrator for Docker Compose stacks
 * OpenAPI spec version: 0.1.0
 */

export interface DeployOutputBody {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** App name derived from image_tag (last path segment before the colon) */
  app: string;
  /** Error message present only when status is failed */
  error?: string;
  /** Unique job identifier (UUID) */
  job_id: string;
  /** Job status: queued, running, success, or failed */
  status: string;
}