# apps/dashboard/src/api/generated/models/errorDetail.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * RollHook API
 * Webhook-driven rolling deployment orchestrator for Docker Compose stacks
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorDetail {
  /** Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id' */
  location?: string;
  /** Error message text */
  message?: string;
  /** The value at the given location */
  value?: unknown;
}