# apps/dashboard/src/api/generated/models/job.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * RollHook API
 * Webhook-driven rolling deployment orchestrator for Docker Compose stacks
 * OpenAPI spec version: 0.1.0
 */

export interface Job {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  app: string;
  compose_path?: string;
  created_at: string;
  error?: string;
  id: string;
  image_tag: string;
  service?: string;
  status: string;
  updated_at: string;
}